+++ src/main/java/buttplugbot/telegrambot/dao/PlugDao.java
@@ -116,9 +116,9 @@ public class PlugDao {
 		storeDB();
 	}
 
-	public String addTemporary(Plug plug, int hours) {
+	public String addTemporary(Plug plug, double hours) {
 		final long now = System.currentTimeMillis();
-		final TemporaryPlug temporaryPlug = new TemporaryPlug(plug, now + hours * 3600000);
+		final TemporaryPlug temporaryPlug = new TemporaryPlug(plug, now + ((long) (hours * 3600000)));
 		final String id = Long.toHexString(createId());
 		temporaryPlugs.put(id, temporaryPlug);
 		return id;
